Move auth page layout into separate component
