+++ ui/src/components/layouts/MapLayout.tsx
@@ -11,9 +11,9 @@ import Header from "@/components/Header";
 import { useFeedQuery, useFeedsQuery } from "@/graphql/generated";
 import { displayDesktopOnly, displayMobileOnly } from "@/styles/responsive";
 
-import Player, { PlayerSpacer } from "./Player";
+import Player, { PlayerSpacer } from "../Player";
 
-const MapWithNoSSR = dynamic(() => import("./Map"), {
+const MapWithNoSSR = dynamic(() => import("../Map"), {
   ssr: false,
 });
